Configura env vars e config maps

+++ song_recommendation_model_generator.py
@@ -1,9 +1,9 @@
 import pandas as pd
+import os
+import argparse
 import pickle
 from fpgrowth_py import fpgrowth
 
-DATASET_1 = "/app/datasets/2023_spotify_ds1.csv"
-DATASET_2 = "/app/datasets/2023_spotify_ds2.csv"
 MODEL_FILE = "/app/datasets/recommendation_model.pickle"
 
 def load_data(file_path):
@@ -15,32 +15,34 @@ def generate_rules(transactions, min_sup=0.1, min_conf=0.3):
     freq_item_set, rules = fpgrowth(transactions, minSupRatio=min_sup, minConf=min_conf)
     return freq_item_set, rules
 
-print(f"Loading data from {DATASET_1}...")
-transactions_ds1 = load_data(DATASET_1)
+def main(dataset_path):
+    print(f"Dataset Path: {dataset_path}")
 
-print("Training model with dataset 1...")
-freq_item_set_ds1, rules_ds1 = generate_rules(transactions_ds1)
+    if not os.path.exists(dataset_path):
+        print(f"Error: Dataset path '{dataset_path}' does not exist.")
+        return
 
-model = {
-    "freq_item_set": freq_item_set_ds1,
-    "rules": rules_ds1,
-    "track_popularity": {}, 
-}
-print(f"Initial model trained with {len(rules_ds1)} rules.")
+    print(f"Loading data from {dataset_path}...")
+    transactions_ds = load_data(dataset_path)
 
-if DATASET_2:
-    print(f"Loading data from {DATASET_2}...")
-    transactions_ds2 = load_data(DATASET_2)
-    
-    print("Updating model with dataset 2...")
-    freq_item_set_ds2, rules_ds2 = generate_rules(transactions_ds2)
-    
-    model["rules"].extend(rules_ds2)
-    model["freq_item_set"].extend(freq_item_set_ds2)
-    print(f"Updated model with {len(rules_ds2)} additional rules.")
+    print("Training model with dataset...")
+    freq_item_set_ds, rules_ds = generate_rules(transactions_ds)
 
-print(f"Saving model to {MODEL_FILE}...")
-with open(MODEL_FILE, "wb") as model_file:
-    pickle.dump(model, model_file)
+    model = {
+        "freq_item_set": freq_item_set_ds,
+        "rules": rules_ds,
+        "track_popularity": {}, 
+    }
+    print(f"Model trained with {len(rules_ds)} rules.")
 
-print(f"Model saved! Total rules: {len(model['rules'])}")
+    print(f"Saving model to {MODEL_FILE}...")
+    with open(MODEL_FILE, "wb") as model_file:
+        pickle.dump(model, model_file)
+
+    print(f"Model saved! Total rules: {len(model['rules'])}")
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--dataset-path', type=str, required=True, help="Path to the dataset")
+    args = parser.parse_args()
+    main(args.dataset_path)
